+++ changelog.md
@@ -2,11 +2,27 @@
 
 📝 following beta format X.Y.Z where Y = breaking change and Z = feature and fix. Later => FAIL.FEATURE.FIX
 
-## 0.9.15(2023-02-04)
+## 0.9.19(2023-03-17)
+
+- Feat: JSON storage
+
+## 0.9.18(2023-03-13)
+
+- Fix: Fix false duplicate $id error
+
+## 0.9.17(2023-03-13)
+
+- Feat: %vars can be used to be consumed in pre-hooks
+
+## 0.9.16(2023-03-04)
+
+- Feat: dependency checker
+
+## 0.9.15(2023-03-04)
 
 - Fix: removed console.log
 
-## 0.9.14(2023-02-03)
+## 0.9.14(2023-03-03)
 
 - Feat: triggers are optional, if none is created, then all are triggered (the $op can be used instead)
 - Fix: Extended objects inherit as well the prehooks. Actions are still unsafe
@@ -47,7 +63,7 @@
 - Feat: Can add children nodes on transformations
 - Feat: Replace is now a valid operation
 - Breaking change: Should use $thing instead of $entity and $relation;
-- Breaking change: Mutation result are always array (even if nothing is nested) 
+- Breaking change: Mutation result are always array (even if nothing is nested)
 - Refacto: Preparing for multiple DBs
 - Chore: Immer downgraded to 9.0.21 (until <https://github.com/immerjs/immer/issues/1087> is fixed)
 - Fix: lot of fixes
@@ -55,7 +71,6 @@
 ## 0.8.11(2023-02-15)
 
 - Tests: Added nested object validation test
-
 
 ## 0.8.10(2023-02-14)